Create Salary
The head of the company notices that more and more employees spend time on sites that distract them.
To prevent this, he introduces surprise checks in the open tabs of his employees' browsers.
According to the open site, the following fines are imposed in the section:
Facebook - 150; Instagram - 100; Reddit - 50.
If during the inspection the salary becomes less than or equal to BGN 0, "You have lost your salary" is displayed on the console. and the program ends. Otherwise, after checking the console, the rest of the salary is displayed.