Create Back to the past
Ivancho is 18 years old and receives an inheritance, which consists of X amount of money and a time machine. He decides to return by 1800, but does not know if the money will be enough to live without work. Write a program that calculates whether Ivancho will have enough money to not have to work until a year inclusive. Assuming that for each even year (1800, 1802, etc.) he will spend BGN 12,000. For each odd (1801, 1803, etc.) he will spend 12,000 + 50 * [the years he has completed in the given year].

Entrance
The input is readable from the console and contains exactly 2 lines:
Inherited money - real number in the interval [1.00 ... 1 000 000.00]
The year he must live (inclusive) - an integer in the range [1801 ... 1900]

Exit
To print on the console 1 line. The amount must be formatted up to two decimal places:
If the money is enough: "Yes! He will live a carefree life and will have {N} dollars left." - where N is the money he will have left.
If the money is NOT enough: "He will need {M} dollars to survive." - where M is the amount that is NOT enough.