Create Hospital
For a period of time, patients come to the hospital every day for an examination. It initially has 7 doctors. Each doctor can see only one patient a day, but sometimes there is a shortage of doctors, so the rest of the patients are sent to other hospitals. Every third day, the hospital makes calculations and if the number of patients not examined is greater than the number of examined, another doctor is appointed. As the appointment is made before the admission of patients for the day.
Write a program that calculates the number of patients examined and not examined for a given period.

Entrance
The input is readable from the console and contains:
In the first row - the period for which you need to make calculations. Integer in the interval [1 ... 1000]
In the following lines (equal to the number of days) - the number of patients who arrive for examination for the current day. Integer in the interval [0… 10 000]

Exit
To print 2 lines on the console:
In the first line: "Treated patients: {number of patients examined}."
On the second line: "Untreated patients: {number of patients not examined}."