Create Logistics
You are responsible for the logistics of various cargoes. Depending on the weight of the load, a different vehicle is required. The price per tonne for which the goods are transported is different for each vehicle:
Up to 3 tons - minibus (BGN 200 per ton)
From 4 to 11 tons - truck (175 BGN per ton)
12 and more tons - train (120 BGN per ton)

Your task is to calculate the average price per tonne of cargo, as well as the percentage of tons transported by each vehicle, relative to the total weight (in tons) of the entire cargo.

Entrance
A series of numbers are read from the console, each on a separate line:
In the first row - the number of goods for transport - an integer in the range [1 ... 1000]
For each load on a separate line - the tonnage of the load - an integer in the range [1 ... 1000]

Get out
Print 4 rows of the console as follows:
First line - the average price per tonne of cargo (rounded to the second decimal place);
Second line - the percentage of tonnes transported by minibus (percentage between 0.00% and 100.00%);
Third row - the percentage of tonnes transported by truck (percentage between 0.00% and 100.00%);
Fourth row - the percentage of tonnes transported by train (percentage between 0.00% and 100.00%).