Create Grades
Write a program to calculate exam evaluation statistics. At the beginning, the program receives the number of students who have passed the exam and for each student their grade. At the end, the program must print the percentage of students with a grade between 2.00 and 2.99, between 3.00 and 3.99, between 4.00 and 4.99, 5.00 or more. Also the average success of the exam.

Entrance
A series of numbers are read from the console, each on a separate line:
In the first row - the number of students who took the exam - an integer in the range [1 ... 1000]
For each student on a separate line - the grade from the exam - a real number in the interval [2.00 ... 6.00]

Get out
Print 5 rows of the console that contain the following information:
Line 1 - "Best students: {percentage of students with a grade of 5.00 or more}%"
Row 2 - "Between 4.00 and 4.99: {between 4.00 and 4.99 inclusive}%"
Row 3 - "Between 3.00 and 3.99: {between 3.00 and 3.99 inclusive}%"
Line 4 - "Failed: {less than 3.00}%"
Line 5 - "Average: {average success}"

All numbers must be formatted to the second decimal place.