Create Game of intervals
Write a program to calculate the result of the game. First you get a number that shows how many moves the game will last. Then for each move of the game you will receive a new number. The points are added according to the interval in which the number falls. If the number is negative or greater than 50, it is invalid. At the beginning of the game the score is 0 and points are added to each move as follows:
 
From 0 to 9: 20% of the number
From 10 to 19: 30% of the number
From 20 to 29: 40% of the number
From 30 to 39:50 points
From 40 to 50: 100 points
Invalid number: the result is divisible by 2
 
In addition to the result, the program must display statistics for percentages of numbers in the given intervals.
Entrance

The input is read from the console:
First line - how many moves will there be during the game - an integer in the interval [1 ... 100]
For each move - the numbers that are checked in which interval they are - integers in the interval [-100 ... 100]

Get out
To print 7 rows to the console:
1st line: "{Final score}"
2nd line: "0 to 9: {percentage in interval}%"
3rd line: "From 10 to 19: {percentage in interval}%"
4th line: "From 20 to 29: {percentage in interval}%"
Line 5: "30 to 39: {percentage in interval}%"
Line 6: "From 40 to 50: {percentage in interval}%"
Line 7: "Invalid numbers: {percentage in interval}%"

All numbers must be formatted to the second decimal place.