Create Football league
The SoftUni team organizes a football tournament.
Initially, we read from the console the capacity of the stadium and the number of all fans. Then, for each fan, read the sector in which it is located. The fans of the first team are in sectors A and B, and of the second - C and D. To write a program that calculates the percentages of fans in each sector, relative to the total number of fans in the stadium, as well as the total percentage of fans for both teams, according to the capacity of the stadium. The total number of fans does NOT exceed the capacity of the stadium.

Entrance
A series of numbers are read from the console, each on a separate line:
1. Capacity of the stadium - an integer in the interval [1… 10000];
2. The number of all fans - an integer in the range [1… 10000].

For each fan, read on a separate line:
The sector in which it is located - text - "A", "B", "V" and "G".

Get out
Print 5 console lines, each containing a percentage between 0.00% and 100.00%, formatted to the second digit after the decimal point:
1. The percentage of fans in sector A
2. The percentage of fans in sector B
3. The percentage of fans in sector B
4. The percentage of fans in sector D
5. The percentage of all fans in relation to the capacity of the stadium.